+++ helpers/index.js
@@ -1,4 +1,7 @@
 const HttpError = require("./HttpError");
+const ctrlWrapper = require("./ctrlWrapper");
+
 module.exports = {
   HttpError,
+  ctrlWrapper,
 };
