+++ middlewares/index.js
@@ -1,0 +1,4 @@
+const validateBody = require("./validateBody");
+module.exports = {
+  validateBody,
+};
