+++ routes/api/contacts.js
@@ -1,91 +1,17 @@
-const contacts = require("../../models/contacts");
 const express = require("express");
-const { HttpError } = require("../../helpers");
-const Joi = require("joi");
-
+const ctrl = require("../../controllers/contacts");
+const { validateBody } = require("../../middlewares");
+const schemas = require("../../schemas/contacts");
 const router = express.Router();
-const addSchema = Joi.object({
-  name: Joi.string().alphanum().min(3).max(30).required(),
-  email: Joi.string()
-    .email({ minDomainSegments: 2, tlds: { allow: ["com", "net"] } })
-    .required(),
-  phone: Joi.string().required(),
-});
 
-const addSchemaForUpdate = Joi.object({
-  name: Joi.string().alphanum().min(3).max(30),
-  email: Joi.string().email({
-    minDomainSegments: 2,
-    tlds: { allow: ["com", "net"] },
-  }),
-  phone: Joi.string(),
-});
+router.get("/", ctrl.getAll);
 
-router.get("/", async (req, res, next) => {
-  try {
-    const allContacts = await contacts.listContacts();
-    res.json(allContacts);
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/:id", ctrl.getById);
 
-router.get("/:id", async (req, res, next) => {
-  try {
-    const { id } = req.params;
-    const contactById = await contacts.getContactById(id);
-    if (!contactById) {
-      throw HttpError(404, "Not found");
-    }
-    res.json(contactById);
-  } catch (error) {
-    next(error);
-  }
-});
+router.post("/", validateBody(schemas.addSchema), ctrl.add);
 
-router.post("/", async (req, res, next) => {
-  try {
-    const { error } = addSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-    const oneContact = await contacts.addContact(req.body);
-    res.status(201).json(oneContact);
-  } catch (error) {
-    next(error);
-  }
-});
+router.put("/:id", validateBody(schemas.addSchemaForUpdate), ctrl.updateById);
 
-router.put("/:id", async (req, res, next) => {
-  try {
-    const { error } = addSchemaForUpdate.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-    const { id } = req.params;
-    const updateContact = await contacts.updateById(id, req.body);
-    if (!updateContact) {
-      throw HttpError(404, "Not found");
-    }
-    res.json(updateContact);
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.delete("/:id", async (req, res, next) => {
-  try {
-    const { id } = req.params;
-    const deleteContact = await contacts.removeContact(id);
-    if (!deleteContact) {
-      throw HttpError(404, "Not found");
-    }
-    res.json({
-      message: "contact deleted",
-    });
-  } catch (error) {
-    next(error);
-  }
-});
+router.delete("/:id", ctrl.removeById);
 
 module.exports = router;
