+++ schemas/contacts.js
@@ -1,0 +1,22 @@
+const Joi = require("joi");
+const addSchema = Joi.object({
+  name: Joi.string().alphanum().min(3).max(30).required(),
+  email: Joi.string()
+    .email({ minDomainSegments: 2, tlds: { allow: ["com", "net"] } })
+    .required(),
+  phone: Joi.string().required(),
+});
+
+const addSchemaForUpdate = Joi.object({
+  name: Joi.string().alphanum().min(3).max(30),
+  email: Joi.string().email({
+    minDomainSegments: 2,
+    tlds: { allow: ["com", "net"] },
+  }),
+  phone: Joi.string(),
+});
+
+module.exports = {
+  addSchema,
+  addSchemaForUpdate,
+};
